refactor(chat): Improve consistency of chat processor
This commit refactors the chat processing system to be more modular and consistent. The checks for newline characters and maximum message length were moved from `chatProcessor.js` into their own dedicated detection modules.

This change centralizes all content-based chat checks into a single, iterable array within the chat processor, improving code readability and maintainability.

Changes include:
- Creating new detection modules: `checkNewline.js` and `checkMaxLength.js`.
- Updating `chatProcessor.js` to use these new modules in its main check array.
- Removing the now-redundant logic directly from `chatProcessor.js`.
- Fixing all associated linting errors that arose from the refactor.
